test(models): keep invoke correction fixtures within budgets
