+++ pkg/services/models/transports/cli/output_required_test.go
@@ -11,6 +11,7 @@ import (
 
 	modelinference "github.com/portpowered/infinite-you/pkg/services/models"
 	"github.com/portpowered/infinite-you/pkg/transports/cli/resolvedinput"
+	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 	"go.uber.org/zap"
 )
 
@@ -934,5 +935,33 @@ func assertInvokeDependencyValues(t *testing.T, cfg InvokeConfig, logger *zap.Lo
 	t.Helper()
 	if cfg.FactoryDir != "" || cfg.HomeDir != "/home/tester" || cfg.Logger != logger || cfg.Diagnostics != diagnostics {
 		t.Fatalf("InvokeConfig dependencies = %#v", cfg)
+	}
+}
+
+func TestAssetPreflightInvocationErrorPreservesOfflineMissingArtifacts(t *testing.T) {
+	t.Parallel()
+
+	cause := &modelinference.AssetOfflineError{Missing: []string{"model.bin", "backend.bin"}}
+	failure := assetPreflightInvocationError("llm", "OMNI", cause)
+	var invocationFailure *modelinference.InvocationFailure
+	if !errors.As(failure, &invocationFailure) || invocationFailure == nil {
+		t.Fatalf("asset preflight failure = %v, want typed invocation failure", failure)
+	}
+	wantMessage := "required model assets are unavailable offline; missing artifacts: backend.bin, model.bin"
+	if invocationFailure.Message != wantMessage {
+		t.Fatalf("offline invocation message = %q, want %q", invocationFailure.Message, wantMessage)
+	}
+	mapped := mapModelsClientError(failure)
+	coded, ok := mapped.(interface {
+		CLIErrorCode() string
+		CLIErrorFamily() factoryapi.ErrorFamily
+		CLIErrorMessage() string
+	})
+	if !ok || coded.CLIErrorCode() != "MODEL_OFFLINE_CACHE_UNAVAILABLE" ||
+		coded.CLIErrorFamily() != factoryapi.ErrorFamilyConflict || coded.CLIErrorMessage() != wantMessage {
+		t.Fatalf("mapped offline failure = %#v, want conflict with complete missing set", mapped)
+	}
+	if !errors.Is(mapped, cause) {
+		t.Fatalf("mapped offline failure = %v, want original AssetOfflineError cause", mapped)
 	}
 }
